refactor: refactor cultural-center
